// src/app/shared/interfaces/user-list.interface.ts
export interface UserList {
  employeeId: number
  username: string
  firstName: string
  lastName: string
  fullName: string
  prefix: string
  email: string
  phoneNumber: string
  employeeCode: string
  createDate: string
  roleCode: string
  position: string
  employeeShift: string
  employeeStatus: string
  buslinesId: number
  busTerminalId: number
  userType: string
  id: number
  platform: string
}